Changed package structure of test classes so that Controller/Service tests are in separate classes/packages
